Create README.md
These are the extended data files for the article entitled: What is it about general practitioner (GP) involved interventions to reduce adverse drug events (ADE) in community dwelling older adults (CDOA) that could inform the development of a successful intervention in an Irish setting; a rapid realist review
Authors: C Waldron, J Hughes, E Wallace, C Cahir, K Bennett
Abstract
Introduction
Older adults in Ireland are at increased risk of adverse drug events (ADE) due, in part, to increasing rates of polypharmacy. Interventions to reduce ADE in community dwelling older adults (CDOA) have had limited success; new approaches are required.
A realist review uses a different lens to examine why and how interventions were supposed to work rather than if, they worked. A rapid realist review (RRR) is a more focused and accelerated version. 
The aim of this RRR is to identify and examine the contexts and mechanisms that play a role in the outcomes relevant to reducing ADE in CDOA in the GP setting that could inform the development of an Irish intervention.
Methodology
Six candidate theories (CT) were developed, based on knowledge of the field and a brief search of recent literature, in relation to how interventions are expected to work. These formed the search strategy. Eighty full texts from 633 abstracts were reviewed, of which 27 were included. Snowballing added a further five articles, a selection of policy documents relevant to the review brought the total articles to 45. Data were extracted relevant to the theories under a series of iteratively developed sub-themes using NVivo software. 
Results
Of the six theories, there was insufficient data for two; a third was rejected as existing barriers in the Irish settings made it impractical. The three remaining theories, relating to GP engagement in interventions, policy documents, and shared decision-making, provided data to guide future interventions to reduce ADEs for CDOA in an Irish setting.
Conclusions
This review provided additional insight in relation to the success or failure of past interventions to reduce ADES for CDOA, specifically what works for whom and under what circumstances, how and why. Future research is required to test our theories within a newly developed intervention.
Keywords Adverse drug events, Community dwelling, Older adult, General Practice